+++ src/voronoi.rs
@@ -17,7 +17,7 @@ impl Voronoi {
         Voronoi { cache,world_seed }
     }
     pub fn get_fuzzy_positions(&mut self, x: i32, y: i32, z: i32) -> (i32, i32, i32) {
-        let key:u128= ((x as u128) << 64 | (y as u128) << 32 | (z as u128)) as u128;
+        let key: u128 = (((x as u32) as u128) << 64 | ((y as u32) as u128) << 32 | ((z as u32) as u128)) as u128;
         let mut value:u128=*self.cache.get(&key).unwrap_or(&u128::MAX);
         if value !=u128::MAX {
             let x:i32= (value >> 64 & 0xFFFF_FFFFu128) as i32;
